Replace unsigned long with uint32_t

+++ test_rclcpp/test/test_intra_process.cpp
@@ -33,7 +33,7 @@ TEST(test_intra_process_within_one_node, nominal_usage) {
   auto publisher = node->create_publisher<test_rclcpp::msg::UInt32>(
     "test_intra_process", custom_qos_profile);
 
-  unsigned long counter = 0;
+  uint32_t counter = 0;
   auto callback =
     [&counter](
     const test_rclcpp::msg::UInt32::SharedPtr msg,
@@ -41,7 +41,7 @@ TEST(test_intra_process_within_one_node, nominal_usage) {
     ) -> void
     {
       ++counter;
-      printf("  callback() %lu with message data %u\n", counter, msg->data);
+      printf("  callback() %4u with message data %u\n", counter, msg->data);
       ASSERT_EQ(counter, msg->data);
       ASSERT_TRUE(message_info.from_intra_process);
     };
