+++ test_rclcpp/test/test_subscription.cpp
@@ -39,12 +39,12 @@ TEST(CLASSNAME(test_subscription, RMW_IMPLEMENTATION), subscription_and_spinning
   auto publisher = node->create_publisher<test_rclcpp::msg::UInt32>(
     "test_subscription", custom_qos_profile);
 
-  unsigned long counter = 0;
+  uint32_t counter = 0;
   auto callback =
     [&counter](const test_rclcpp::msg::UInt32::SharedPtr msg) -> void
     {
       ++counter;
-      printf("  callback() %lu with message data %u\n", counter, msg->data);
+      printf("  callback() %4u with message data %u\n", counter, msg->data);
       ASSERT_EQ(counter, msg->data);
     };
 
@@ -148,12 +148,12 @@ TEST(CLASSNAME(test_subscription, RMW_IMPLEMENTATION), subscription_shared_ptr_c
   auto publisher = node->create_publisher<test_rclcpp::msg::UInt32>(
     "test_subscription", rmw_qos_profile_default);
 
-  unsigned long counter = 0;
+  uint32_t counter = 0;
   auto callback =
     [&counter](test_rclcpp::msg::UInt32::ConstSharedPtr msg) -> void
     {
       ++counter;
-      printf("  callback() %lu with message data %u\n", counter, msg->data);
+      printf("  callback() %4u with message data %u\n", counter, msg->data);
       ASSERT_EQ(counter, msg->data);
     };
 
@@ -182,7 +182,7 @@ TEST(CLASSNAME(test_subscription, RMW_IMPLEMENTATION), subscription_shared_ptr_c
 
   executor.spin_node_some(node);
   // spin up to 4 times with a 25 ms wait in between
-  for (unsigned long i = 0; i < 4 && counter == 0; ++i) {
+  for (uint32_t i = 0; i < 4 && counter == 0; ++i) {
     printf("callback not called, sleeping and trying again\n");
     std::this_thread::sleep_for(std::chrono::milliseconds(25));
     executor.spin_node_some(node);
@@ -196,13 +196,13 @@ TEST(CLASSNAME(test_subscription, RMW_IMPLEMENTATION), subscription_shared_ptr_c
   auto publisher = node->create_publisher<test_rclcpp::msg::UInt32>(
     "test_subscription", rmw_qos_profile_default);
 
-  unsigned long counter = 0;
+  uint32_t counter = 0;
   auto callback =
     [&counter](test_rclcpp::msg::UInt32::ConstSharedPtr msg,
       const rmw_message_info_t & info) -> void
     {
       ++counter;
-      printf("  callback() %lu with message data %u\n", counter, msg->data);
+      printf("  callback() %4u with message data %u\n", counter, msg->data);
       ASSERT_EQ(counter, msg->data);
       ASSERT_FALSE(info.from_intra_process);
     };
@@ -230,7 +230,7 @@ TEST(CLASSNAME(test_subscription, RMW_IMPLEMENTATION), subscription_shared_ptr_c
 
   executor.spin_node_some(node);
   // spin up to 4 times with a 25 ms wait in between
-  for (unsigned long i = 0; i < 4 && counter == 0; ++i) {
+  for (uint32_t i = 0; i < 4 && counter == 0; ++i) {
     printf("callback not called, sleeping and trying again\n");
     std::this_thread::sleep_for(std::chrono::milliseconds(25));
     executor.spin_node_some(node);
